fallback: Consider all Boot* vars when checking for duplicates
Some firmware implementations like the one on the Acer TravelMate
P449-G2-MG completely ignore the contents of BootOrder on boot, and
overwrite it with a value of its own. On this particular machine, the
boot entry that was just created by fallback on the previous boot is not
included by the firmware on this new BootOrder, so it is not considered
when checking for duplicates.

This problem is agravated by the fact that the aformentioned firmware
does not give the user the possibility to boot from the entry created by
fallback (or any other entry created by the OS). The only way to boot
a distro that deploys the fallback setup (no grub inside \EFI\BOOT) with
this firmware is to select the entry pointing to \EFI\BOOT\bootx64.efi,
leading to a new boot entry being created by fallback on every boot.

This commit makes fallback try every Boot* variable when checking for
duplicates, working around this problem.

Signed-off-by: João Paulo Rechi Vita <jprvita@endlessm.com>

https://phabricator.endlessm.com/T15481