Enable signing of fallback and MokManager
Tells the build-system to sign fallback and MokManager with sbsign and
install the signed variants to the ESP.

Signing fallback and MokManager with a throwaway key used to be the
default behavior, but it is now controlled by whether ENABLE_SHIM_CERT
is defined. And now selecting sbsign through the build system is
supported upstream, so we can drop the downstream changes we chand
swapping pesign with sbsign.

https://phabricator.endlessm.com/T24681