[Endless] fallback: Clean-up duplicate boot entries
Consider existing entries with the same label we are trying to create as
duplicates. When a duplicate is detected, simply remove the boot entry
and its occurrence in BootOrder, and let the rest of the logic work as
if there was no boot entry.

This change breaks setups with more than one storage device where there
is a boot entry with the same label on a different device -- for
example, if the user installed the same OS on two different disks, each
with its own bootloader and boot entry. We don't support such setups
with Endless OS.

We need this change for Endless OS because we ship images that are
already partitioned, so we need to modify the partition UUID for all
partitions on first boot otherwise every installation of the same
release would have the same partition UUIDs. With this change we can
rely on fallback to create a new boot entry pointing to the new ESP UUID
on the second boot.

https://phabricator.endlessm.com/T14430
https://phabricator.endlessm.com/T16731