[Endless] Partially revert "fallback: work around the issue of boot option creation with AMI BIOS"
This partially reverts commit 0cc030c2f2fba53b74fb09466a07b8e6297a52d3.

We are going to treat any entries where the label is the same as the one
we are trying to create as duplicates, so we don't need to use
check_masked_boot_option to compare device paths patched with
ami_masked_device_path_guid. But we still want to make sure our
candidate buffer is large enough to store these kind of entries, as we
still want to compare their labels with the label we are trying to
create.

https://phabricator.endlessm.com/T24681