fallback: Consider all Boot* vars when checking for duplicates
Some firmware implementations like the one on the Acer TravelMate
P449-G2-MG completely ignore the value of BootOrder that was previously
set, and overwrite it with a value of its own. On this particular
machine, the boot entry that was just created by fallback on the
previous boot is not included by the firmware on this new BootOrder, so
it is not considered when checking for duplicates.

This problem is agravated by the fact that the aformentioned firmware
does not give the user the possibility to manually boot from any entry
created outside of the firmware setup program. The only way to boot a
distro that deploys the fallback setup and no \EFI\BOOT\grubx64.efi with
this firmware is through \EFI\BOOT\bootx64.efi. This has a side effect
of having a new boot entry being created by fallback on every boot.

This commit makes fallback try every Boot* variable when checking for
duplicates, so at least if finds the duplicate Boot entry and re-uses
it, even if it not listed on BootOrder.

https://phabricator.endlessm.com/T15481