[Endless] fallback: Always try to chain-load the first entry
Some Acer firmwares do not respect the BootOrder variable set by the
system, so unless the user manually creates a boot entry in the BIOS
setup the firmware will always boot the \EFI\BOOT\BOOTX64.EFI path.

In the past this would almost always work as fallback would always try
to chain-load the newly created entry. But after commit 431b8a2e75 when
a TPM is detected on the system, the machine is rebooted instead, and on
the next boot \EFI\BOOT\BOOTX64.EFI gets booted again and the system
ends up in an infinite reboot loop.

This commit reverts back to the previous behavior of always trying to
chain-load the first boot entry and only rebooting when that fails.

https://phabricator.endlessm.com/T24681