[Endless] fallback: Do not remove vars while listing var names
The EFI variable for a duplicate boot entry cannot be removed in the
same loop listing variables from the firmware, otherwise the next call
to GetNextVariableName will fail with EFI_INVALID_PARAMETER because the
variable name being passed will not exist anymore.

This commit changes the logic to save the variables that should be
removed to a list, and then traverses the list removing the variables
after listing all variables from the firmware.

Fixes commit "[Endless] fallback: Clean-up duplicate boot entries".

https://phabricator.endlessm.com/T14430
https://phabricator.endlessm.com/T16731
https://phabricator.endlessm.com/T24681